Factor out new mod util.
