+++ oak_remote_attestation_verification/src/endorsement.rs
@@ -16,56 +16,16 @@
 
 //! Verifies binary endorsements as coming from Transparent Release.
 
-use alloc::{string::String, vec::Vec};
-
-use crate::rekor::{equal_keys, get_rekor_log_entry_body, verify_rekor_log_entry};
 use anyhow::Context;
 use base64::{prelude::BASE64_STANDARD, Engine as _};
 use oak_transparency_claims::claims::{
     parse_endorsement_statement, validate_endorsement, verify_validity_duration,
 };
 
-/// Reference values used by the verifier to appraise the attestation evidence.
-/// <https://www.rfc-editor.org/rfc/rfc9334.html#name-reference-values>
-pub struct ReferenceValue {
-    pub binary_hash: Vec<u8>,
-}
-
-const PEM_HEADER: &str = "-----BEGIN PUBLIC KEY-----";
-const PEM_FOOTER: &str = "-----END PUBLIC KEY-----";
-
-/// Makes a plausible guess whether the public key is in PEM format.
-pub fn looks_like_pem(maybe_pem: &str) -> bool {
-    let p = maybe_pem.trim();
-    p.starts_with(PEM_HEADER) && p.ends_with(PEM_FOOTER)
-}
-
-/// Converts a pem key to raw. Will panic if it does not like like pem.
-pub fn convert_pem_to_raw(public_key_pem: &str) -> anyhow::Result<Vec<u8>> {
-    let stripped = public_key_pem
-        .trim()
-        .strip_prefix(PEM_HEADER)
-        .expect("could not find expected header")
-        .strip_suffix(PEM_FOOTER)
-        .expect("could not find expected footer");
-    let remove_newlines = stripped.replace('\n', "");
-
-    Ok(BASE64_STANDARD.decode(remove_newlines)?)
-}
-
-pub fn convert_raw_to_pem(public_key: &[u8]) -> String {
-    let mut pem = String::from(PEM_HEADER);
-    for (i, c) in BASE64_STANDARD.encode(public_key).chars().enumerate() {
-        if i % 64 == 0 {
-            pem.push('\n');
-        }
-        pem.push(c);
-    }
-    pem.push('\n');
-    pem.push_str(PEM_FOOTER);
-    pem.push('\n');
-    pem
-}
+use crate::{
+    rekor::{get_rekor_log_entry_body, verify_rekor_log_entry},
+    util::{convert_pem_to_raw, equal_keys},
+};
 
 /// Verifies the binary endorsement for a given measurement.
 pub fn verify_binary_endorsement(
